+++ packages/manageState/src/index.ts
@@ -1,37 +1,37 @@
 import { OnRpcRequestHandler } from '@metamask/snaps-types';
 
 export const onRpcRequest: OnRpcRequestHandler = async ({ request }) => {
-  let state = (await snap.request({
+  let state = (await wallet.request({
     method: 'snap_manageState',
-    params: { operation: 'get' },
+    params: ['get'],
   })) as { testState: string[] } | null;
 
   if (!state) {
     state = { testState: [] };
     // initialize state if empty and set default data
-    await snap.request({
+    await wallet.request({
       method: 'snap_manageState',
-      params: { operation: 'update', newState: state },
+      params: ['update', state],
     });
   }
 
   switch (request.method) {
     case 'storeTestData':
       state.testState.push((request.params as string[])[0]);
-      await snap.request({
+      await wallet.request({
         method: 'snap_manageState',
-        params: { operation: 'update', newState: state },
+        params: ['update', state],
       });
       return true;
     case 'retrieveTestData':
-      return await snap.request({
+      return await wallet.request({
         method: 'snap_manageState',
-        params: { operation: 'get' },
+        params: ['get'],
       });
     case 'clearTestData':
-      await snap.request({
+      await wallet.request({
         method: 'snap_manageState',
-        params: { operation: 'clear' },
+        params: ['clear'],
       });
       return true;