+++ packages/notification/src/index.ts
@@ -6,20 +6,24 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
 }) => {
   switch (request.method) {
     case 'inApp':
-      return snap.request({
+      return wallet.request({
         method: 'snap_notify',
-        params: {
-          type: 'inApp',
-          message: `TEST INAPP NOTIFICATION`,
-        },
+        params: [
+          {
+            type: 'inApp',
+            message: `TEST INAPP NOTIFICATION`,
+          },
+        ],
       });
     case 'native':
-      return snap.request({
+      return wallet.request({
         method: 'snap_notify',
-        params: {
-          type: 'native',
-          message: `Hello, ${origin}!`,
-        },
+        params: [
+          {
+            type: 'native',
+            message: `Hello, ${origin}!`,
+          },
+        ],
       });
     default:
       throw new Error('Method not found.');