+++ packages/rpc/CHANGELOG.md
@@ -6,9 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.5.0]
+### Changed
+- No changes this release.
+
+## [4.4.1]
+### Changed
+- No changes this release.
+
+## [4.4.0]
+### Changed
+- No changes this release.
+
+## [4.3.0]
+### Changed
+- No changes this release.
+
 ## [4.2.0]
 ### Added
 - Initial release ([#105](https://github.com/MetaMask/test-snaps/pull/105))
 
-[Unreleased]: https://github.com/MetaMask/test-snaps/compare/v4.2.0...HEAD
+[Unreleased]: https://github.com/MetaMask/test-snaps/compare/v4.5.0...HEAD
+[4.5.0]: https://github.com/MetaMask/test-snaps/compare/v4.4.1...v4.5.0
+[4.4.1]: https://github.com/MetaMask/test-snaps/compare/v4.4.0...v4.4.1
+[4.4.0]: https://github.com/MetaMask/test-snaps/compare/v4.3.0...v4.4.0
+[4.3.0]: https://github.com/MetaMask/test-snaps/compare/v4.2.0...v4.3.0
 [4.2.0]: https://github.com/MetaMask/test-snaps/releases/tag/v4.2.0